requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
langchain==0.2.0
langchain-core==0.2.0
langchain-community==0.2.0
langchain-huggingface==0.0.3
langchain-text-splitters==0.2.0
llama-index==0.10.0
llama-index-core==0.10.0
llama-index-embeddings-huggingface==0.1.5
llama-index-readers-file==0.10.0
BCEmbedding==0.1.0
torch==2.1.0
transformers==4.35.0
sentence-transformers==2.2.2
PyPDF2==3.0.1
psycopg2-binary==2.9.9
pgvector==0.2.4
numpy==1.24.3
loguru==0.7.2
pytest==7.4.3
requests==2.31.0
python-multipart==0.0.6
chromadb==0.4.24
faiss-cpu==1.7.4
rank-bm25==0.2.2
